add: src folder. Move 'components', 'hooks' and 'styles' into src folder
