+++ homeworks/edelmiro-anton/21-csr-ssr-ssg-nextjs/avatar-fetch/pages/_app.js
@@ -1,4 +1,4 @@
-import '../styles/globals.css'
+import '../src/styles/globals.css'
 
 function MyApp({ Component, pageProps }) {
   return <Component {...pageProps} />
