+++ homeworks/edelmiro-anton/21-csr-ssr-ssg-nextjs/avatar-fetch/src/components/Layout/Layout.module.css
