+++ homeworks/edelmiro-anton/21-csr-ssr-ssg-nextjs/avatar-fetch/src/components/Loader/Loader.js
