+++ homeworks/edelmiro-anton/21-csr-ssr-ssg-nextjs/avatar-fetch/src/hooks/useAvatars.js
